Gestión de datasets y Gestión de enfermedades
Se realizaron modificaciones en el menú administrador creando un despliegue de los diferentes tipos de usuarios.
Se realizó modificaciones en el código a modo de dejarlo más estético y entendible.
Se realizó la gestión de datasets en donde se descomprime las imágenes y al final se modifica la ruta del dataset haciendo referencia a la carpeta creada donde almacena las imágenes de entrenamiento y validación. Incluye las cuatro operaciones (Crear, Consultar, Actualizar y Eliminar)
Se realizó la gestión de las enfermedades con las cuatro operaciones (Crear, Consultar, Actualizar y Eliminar)

+++ SafeCrops/AppSafeCrops/models.py
@@ -3,6 +3,7 @@ from django.db import models
 from django.contrib.auth.models import User #Libreria para usar la tabla usuarios por defecto de Django
 from django.utils import timezone
 from django.contrib.auth.models import User
+import shutil
 
 # Create your models here.
 
@@ -23,7 +24,7 @@ class Administrador(models.Model): #Se crea el modelo Administrador
     apellidoM = models.CharField(max_length=45, verbose_name='Apellido M', null=True, blank=True, default='') #Se crea el campo apellidoM que es un campo de tipo cadena de caracteres
     fechaNac = models.DateField(auto_now_add=False, auto_now=False, blank=True ,verbose_name='Fecha de Nacimiento', default=timezone.now) #Se crea el campo fechaNac que es un campo de tipo fecha
     correo = models.EmailField(max_length=50 ,verbose_name='Correo', null=True, blank=True, default='') #Se crea el campo correo que es un campo de tipo correo
-    imagen = models.ImageField(max_length=100, upload_to='imagenes/', verbose_name='Imagen', default='imagenes/defaultProfile/defaultProfile.png') #Se crea el campo imagen que es un campo de tipo imagen
+    imagen = models.ImageField(max_length=100, upload_to='imagenes/', verbose_name='Imagen', default='imagenes/defaultProfile.png') #Se crea el campo imagen que es un campo de tipo imagen
     userType = models.CharField(max_length=20, verbose_name='', default='Administrador') #Se crea el campo userType que es un campo de tipo cadena de caracteres
     user = models.OneToOneField(User, on_delete=models.CASCADE, verbose_name='ID_Usuario', related_name='administradorUser', null=True, blank=True) #Se crea el campo user que es una relación uno a uno con el modelo User de django
     
@@ -33,7 +34,9 @@ class Administrador(models.Model): #Se crea el modelo Administrador
         return fila #Se retorna la variable fila
 
     def delete(self, using=None, keep_parents=False): #Se sobreescribe el método delete para que no se elimine el usuario de django
-        self.imagen.storage.delete(self.imagen.name)
+        defaultProfile = str(self.imagen.name).split('/')
+        if defaultProfile[1] != 'defaultProfile.png':
+            self.imagen.storage.delete(self.imagen.name)
         super().delete() #Se llama al método delete de la clase padre
 
 class Experto(models.Model): #Se crea el modelo Experto
@@ -44,7 +47,7 @@ class Experto(models.Model): #Se crea el modelo Experto
     fechaNac = models.DateField(auto_now_add=False, auto_now=False, blank=True ,verbose_name='Fecha de Nacimiento', default=timezone.now) #Se crea el campo fechaNac que es un campo de tipo fecha
     correo = models.EmailField(max_length=50, verbose_name='Correo', null=True) #Se crea el campo correo que es un campo de tipo correo
     institucionPerteneciente = models.CharField(max_length=100, verbose_name='Institución a la que pertenece') #Se crea el campo correo que es un campo de tipo correo
-    imagen = models.ImageField(max_length=100, upload_to='imagenes/', verbose_name='Imagen', default='imagenes/defaultProfile/defaultProfile.png') #Se crea el campo imagen que es un campo de tipo imagen
+    imagen = models.ImageField(max_length=100, upload_to='imagenes/', verbose_name='Imagen', default='imagenes/defaultProfile.png') #Se crea el campo imagen que es un campo de tipo imagen
     userType = models.CharField(max_length=20, verbose_name='', default='Experto') #Se crea el campo userType que es un campo de tipo cadena de caracteres
     user = models.OneToOneField(User, on_delete=models.CASCADE, verbose_name='ID_Usuario', related_name='expertoUser', null=True, blank=True) #Se crea el campo user que es una relación uno a uno con el modelo User de django
 
@@ -53,7 +56,9 @@ class Experto(models.Model): #Se crea el modelo Experto
         return fila #Se retorna la variable fila
 
     def delete(self, using=None, keep_parents=False): #Se sobreescribe el método delete para que no se elimine el usuario de django
-        self.imagen.storage.delete(self.imagen.name)
+        defaultProfile = str(self.imagen.name).split('/')
+        if defaultProfile[1] != 'defaultProfile.png':
+            self.imagen.storage.delete(self.imagen.name)
         super().delete() #Se llama al método delete de la clase padre
 
 class Tester(models.Model): #Se crea el modelo Tester
@@ -63,7 +68,7 @@ class Tester(models.Model): #Se crea el modelo Tester
     apellidoM = models.CharField(max_length=45, verbose_name='Apellido M', null=True, blank=True)  #Se crea el campo apellidoM que es un campo de tipo cadena de caracteres
     fechaNac = models.DateField(auto_now_add=False, auto_now=False, blank=True ,verbose_name='Fecha de Nacimiento', default=timezone.now) #Se crea el campo fechaNac que es un campo de tipo fecha
     correo = models.EmailField(max_length=50, verbose_name='Correo', null=True)    #Se crea el campo correo que es un campo de tipo correo
-    imagen = models.ImageField(max_length=100, upload_to='imagenes/', verbose_name='Imagen', default='imagenes/defaultProfile/defaultProfile.png') #Se crea el campo imagen que es un campo de tipo imagen
+    imagen = models.ImageField(max_length=100, upload_to='imagenes/', verbose_name='Imagen', default='imagenes/defaultProfile.png') #Se crea el campo imagen que es un campo de tipo imagen
     userType = models.CharField(max_length=20, verbose_name='', default='Tester') #Se crea el campo userType que es un campo de tipo cadena de caracteres
     user = models.OneToOneField(User, on_delete=models.CASCADE, verbose_name='ID_Usuario', related_name='testerUser', null=True, blank=True) #Se crea el campo user que es una relación uno a uno con el modelo User
 
@@ -71,6 +76,34 @@ class Tester(models.Model): #Se crea el modelo Tester
         fila = self.nombre
         return fila
 
+    def delete(self, using=None, keep_parents=False): #Se sobreescribe el método delete para que no se elimine el usuario de django
+        defaultProfile = str(self.imagen.name).split('/')
+        if defaultProfile[1] != 'defaultProfile.png':
+            self.imagen.storage.delete(self.imagen.name)
+        super().delete() #Se llama al método delete de la clase padre
+
+class Enfermedad(models.Model):
+    id_Enfermedad = models.AutoField(primary_key=True)
+    nombreEnfermedad = models.CharField(max_length=45, unique=True, verbose_name='Nombre de la Enfermedad')
+    cultivoEnfermedad = models.CharField(max_length=45, verbose_name='Cultivo en el que se presenta la enfermedad')
+    descripcionEnfermedad = models.TextField(max_length=200, verbose_name='Descripción de la enfermedad')
+    curaEnfermedad = models.TextField(max_length=200, verbose_name='Cura de la enfermedad')
+
+class Dataset(models.Model):
+    id_Dataset = models.AutoField(primary_key=True)
+    nombreDataset = models.CharField(max_length=45, unique=True, verbose_name='Nombre del Dataset')
+    ruta = models.FileField(max_length=100, upload_to='datasets/', verbose_name='Ruta del Dataset')
+    numImgTotal = models.IntegerField(verbose_name='Número de imágenes totales')
+    numImgEntrenamiento = models.IntegerField(verbose_name='Número de imágenes de entrenamiento')
+    numImgValidacion = models.IntegerField(verbose_name='Número de imágenes de validación')
+    segmentacion_SAM = models.CharField(max_length=5, verbose_name='Segmentación SAM')
+    formatoImg = models.CharField(max_length=10, verbose_name='Formato de las imágenes')
+    tipoDataset = models.CharField(max_length=45, verbose_name='Tipo de dataset')
+
+    def __str__(self):
+        fila = self.nombreDataset
+        return fila
+
     def delete(self, using=None, keep_parents=False):
-        self.imagen.storage.delete(self.imagen.name)
+        shutil.rmtree(self.ruta.name)
         super().delete()
